Add the filename to Rego fragment locations.
It is awkward to map the location of Rego erros back to the right
line on the test document, but we can improve things (in a hacky way)
by capturing the file and line number in the filename that we feed to
Rego when we compile a fragment.

This updates #12.

Signed-off-by: James Peach <jpeach@vmware.com>